arrumei as paradas do pagamento

+++ gestaoapp/models/pagamentos.py
@@ -7,7 +7,7 @@ from gestaoapp.models.vinculo import Vinculo
 class Pagamentos(models.Model):
     dt_pagamento = models.DateField(null=True, blank=True)
     vinculo = models.ForeignKey(Vinculo)
-    valor = models.DecimalField(max_digits=3, decimal_places=2, null=True)
+    valor = models.FloatField(null=True)
 
     def __str__(self):
         return ('%s %s %s' % (self.dt_pagamento, self.vinculo, self.vinculo.bolsa))
